requirements.txt
amqp==5.1.1
asgiref==3.6.0
async-timeout==4.0.2
billiard==3.6.4.0
cached-property==1.5.2
cachetools==5.3.0
celery==5.2.7
certifi @ file:///private/var/folders/sy/f16zz6x50xz3113nwtb9bvq00000gp/T/abs_477u68wvzm/croot/certifi_1671487773341/work/certifi
charset-normalizer==3.0.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
Django==3.2.18
django-auth-ldap==4.1.0
django-celery-results==2.5.0
django-cors-headers==4.1.0
django-redis==5.2.0
djongo==1.3.6
dnspython==2.3.0
docopt==0.6.2
google-api-core==2.11.0
google-api-python-client==2.73.0
google-auth==2.16.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.8.0
googleapis-common-protos==1.58.0
gunicorn==20.1.0
httplib2==0.21.0
idna==3.4
importlib-metadata==4.13.0
kombu==5.2.4
ldap3==2.9.1
mysqlclient==2.1.1
oauth2client==4.1.3
oauthlib==3.2.2
Pillow==9.4.0
pipreqs==0.4.13
prompt-toolkit==3.0.36
protobuf==4.21.12
psycopg2-binary==2.9.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pymongo==3.12.3
pymssql==2.2.7
pyparsing==3.0.9
python-ldap==3.4.3
pytz==2022.7.1
PyYAML==6.0
redis==4.5.4
requests==2.28.2
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
sqlparse==0.2.4
typing_extensions==4.4.0
uritemplate==4.1.1
urllib3==1.26.14
uWSGI==2.0.21
vine==5.0.0
wcwidth==0.2.6
websockets==10.4
yarg==0.1.9
zipp==3.11.0
